--- client-react/src/pages/app/functions/create/CreateCard.helper.ts
@@ -1,7 +1,0 @@
-import { FunctionTemplate } from '../../../../models/functions/function-template';
-import { PivotState } from './FunctionCreate';
-
-export function onTemplateSelected(functionTemplate: FunctionTemplate, setSelectedFunctionTemplate: any, setPivotStateKey: any) {
-  setSelectedFunctionTemplate(functionTemplate);
-  setPivotStateKey(PivotState.details);
-}
